linkerd-web public-api/viz updates
This is based on the branch `alpeb/linkerd-metrics-api-bootstrap` from #5558.

`linkerd-web` requires connecting with both the public-api and the viz api, so both addresses (and the viz namespace) are now provided as parameters to the container.

We're also updating some dependencies that still were pointing to the public-api and now should point to the viz api.

Tests are not expected to pass on this one either.